+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/json/NetworkModificationDataJsonModule.java
@@ -4,16 +4,19 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.modification.topology.data;
+package com.powsybl.iidm.modification.json;
 
 import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.powsybl.iidm.modification.data.topology.CreateGeneratorBayData;
+import com.powsybl.iidm.modification.json.topology.CreateGeneratorBayDataDeserializer;
+import com.powsybl.iidm.modification.json.topology.CreateGeneratorBayDataSerializer;
 
 /**
  * @author Miora Vedelago <miora.ralambotiana at rte-france.com>
  */
-public class CreateGeneratorBayDataJsonModule extends SimpleModule {
+public class NetworkModificationDataJsonModule extends SimpleModule {
 
-    public CreateGeneratorBayDataJsonModule() {
+    public NetworkModificationDataJsonModule() {
         addSerializer(CreateGeneratorBayData.class, new CreateGeneratorBayDataSerializer());
         addDeserializer(CreateGeneratorBayData.class, new CreateGeneratorBayDataDeserializer());
     }
